backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
apscheduler==3.10.4
PyJWT==2.8.0
gunicorn==21.2.0
